tile/pyproject.toml
[project]
name = "tile"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cuda-tile>=1.0.0",
    "cupy-cuda13x>=13.6.0",
    "numpy>=2.3.5",
]
